Added a method to clean the points returned by get_points.
get_points returns a LOT of points. This is fine if the points are on
a complex curve, but when the points are all on straight line
segments this is unnecessary. A lot of beam functions will
return straight line segments. clean_points walks through the
points and removes the middle one from any group of 3x points that
form a straight line.